==> 1_data_preparation.ipynb <==
https://huggingface.co/docs/transformers/en/model_doc/encoder-decoder

# Dataset

import os
import json
import pprint
import re
import pickle
import time
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm import tqdm
import warnings
warnings.filterwarnings("ignore")
from wordcloud import WordCloud
import json

def load_json_files_from_directory(directory):
    data = []
    for filename in os.listdir(directory):
        if filename.endswith(".json"):
            file_path = os.path.join(directory, filename)
            with open(file_path, 'r', encoding='utf-8') as file:
                data.append(json.load(file))
    return data

folder_data = "E:\\Dataset\\liputan6_data2"
dataset_type = "dataset"

train_dir = os.path.join(folder_data, dataset_type, "train")
valid_dir = os.path.join(folder_data, dataset_type, "dev")
test_dir = os.path.join(folder_data, dataset_type, "test")

train_dir

train_data = load_json_files_from_directory(train_dir)
valid_data = load_json_files_from_directory(valid_dir)
test_data = load_json_files_from_directory(test_dir)

train_data[0]

train_set = pd.json_normalize(train_data)
valid_set = pd.json_normalize(valid_data)
test_set = pd.json_normalize(test_data)

def normalize_text(tokenized_text):
    normalized_text = ""
    for sentence in tokenized_text:
        for word in sentence:
            if word in [".", ",", ":", ";", "!", "?", ")", "]"]:
                normalized_text = normalized_text.rstrip() + word + " "
            elif word in ["(", "["]:
                normalized_text += word
            else:
                normalized_text += word + " "
    return normalized_text.strip()


train_set['combined_clean_article'] = train_set['clean_article'].apply(normalize_text)
train_set['combined_clean_summary'] = train_set['clean_summary'].apply(normalize_text)

train_set.head()

train_set.sample(1)[["combined_clean_article", "combined_clean_summary"]].values

valid_set['combined_clean_article'] = valid_set['clean_article'].apply(normalize_text)
valid_set['combined_clean_summary'] = valid_set['clean_summary'].apply(normalize_text)
test_set['combined_clean_article'] = test_set['clean_article'].apply(normalize_text)
test_set['combined_clean_summary'] = test_set['clean_summary'].apply(normalize_text)

train_set.shape, valid_set.shape, test_set.shape

folder_data = "E:\\Dataset\\liputan6_data2\\"
train_set.to_csv(folder_data+'train_set.csv', index=False)
valid_set.to_csv(folder_data+'valid_set.csv', index=False)
test_set.to_csv(folder_data+'test_set.csv', index=False)
