==> 3_data_preprocessing.ipynb <==
import pandas as pd
import re

folder_data = "E:\\Dataset\\liputan6_data2\\"
train_set = pd.read_csv(folder_data+'train_set.csv')
valid_set = pd.read_csv(folder_data+'valid_set.csv')
test_set = pd.read_csv(folder_data+'test_set.csv')

def clean_article(text):
    text = re.sub(r'Liputan6\. com, [A-Za-z\s]+:', '', text)
    text = re.sub(r'\w+\s*\(\d{1,2}/\d{1,2}\)', '', text)
    text = re.sub(r'\([A-Z]+/[A-Za-z\s]+(?: dan [A-Za-z\s]+)?\)\.', '', text)
    text = re.sub(r'\[baca: .*?\]', '', text)
    return text.strip()

train_set['final_clean_article'] = train_set['combined_clean_article'].apply(clean_article)
valid_set['final_clean_article'] = valid_set['combined_clean_article'].apply(clean_article)
test_set['final_clean_article'] = test_set['combined_clean_article'].apply(clean_article)

final_train_set = train_set[['final_clean_article','combined_clean_summary']]
final_valid_set = valid_set[['final_clean_article','combined_clean_summary']]
final_test_set = test_set[['final_clean_article','combined_clean_summary']]

final_train_set.head()

final_valid_set.head()

final_test_set.head()

final_train_set.shape, final_valid_set.shape, final_test_set.shape

folder_data = "E:\\Dataset\\liputan6_data2\\"
final_test_set.to_csv(folder_data+'final_test_set.csv', index=False)
final_valid_set.to_csv(folder_data+'final_valid_set.csv', index=False)
final_train_set.to_csv(folder_data+'final_train_set.csv', index=False)
